Added cubic spline interpolation feature for IRB140 trajectories.
Whereas before the LCM Monitor would take a raw trajectory plan of keyframes (often 20 frames) and pass those as position commands to the arm, now there is an intermediate step of creating a cubic spline from these keyframes and resampling at a higher rate. This should make for more smooth arm movement.

There still may remain work to be done if we also wish to specify velocities to the arm (presently everything is sent through the position buffer).